Rename files to better describe intent
Also removed redundant files and tests
