+++ server/plugins/router.js
@@ -19,8 +19,7 @@ const routes = [].concat(
   require('../routes/stations-within-target-area'),
   require('../routes/warnings-alerts-within-station-buffer'),
   require('../routes/thresholds-data'),
-  require('../routes/thresholds-data-imtd'),
-  require('../routes/station-threshold'),
+  require('../routes/station-imtd-threshold'),
   require('../routes/stations-health'),
   require('../routes/telemetry-health'),
   require('../routes/ffoi-health'),
